+++ NEWS.md
@@ -1,5 +1,10 @@
 # labelled (development version)
 
+* a `.strict` option has been added to `set_variable_labels()` allowing
+  to pass values for columns not observed in the data (it could be useful
+  for using a same list of labels for several data.frame sharing some 
+  variables).
+
 # labelled 2.6.0
 
 * `look_for()` has been redesigned: